Implemented filters and SUBSCRIBE_DONE (kinda) in the demos

+++ examples/moq-pub-options.h
@@ -11,6 +11,8 @@
 #ifndef MOQ_PUB_OPTIONS
 #define MOQ_PUB_OPTIONS
 
+#include <stdint.h>
+
 #include <glib.h>
 
 /*! \brief Struct containing the parsed command line options */
@@ -18,7 +20,7 @@ typedef struct demo_options {
 	char *moq_version;
 	const char **track_namespace;
 	const char *track_name;
-	int first_group;
+	uint64_t first_group;
 	const char *auth_info;
 	const char *delivery;
 	gboolean extensions;
